Separated basics tests into it's own class
