+++ android-uribeacon/uribeacon-validator/src/main/java/org/uribeacon/validator/UriBeaconTests.java
@@ -27,14 +27,12 @@ import org.uribeacon.validator.TestHelper.TestCallback;
 
 import java.util.ArrayList;
 
-public class UriBeaconTests {
+public class UriBeaconTests{
 
   private static final String TAG = UriBeaconTests.class.getCanonicalName();
   public static ArrayList<TestHelper> initializeTests(Context context, BluetoothDevice bluetoothDevice, TestCallback testCallback, boolean optional) {
 
     ArrayList<TestHelper> tests = new ArrayList<>();
-
-    tests.addAll(basicTests(context, bluetoothDevice, testCallback));
     tests.addAll(dataTests(context, bluetoothDevice, testCallback));
     tests.addAll(flagsTests(context, bluetoothDevice, testCallback));
     if (optional) {
@@ -57,37 +55,6 @@ public class UriBeaconTests {
   ///////////////////
   ////// Tests //////
   ///////////////////
-  private static ArrayList<TestHelper> basicTests(Context context, BluetoothDevice bluetoothDevice,
-      TestCallback testCallback) {
-
-    ArrayList<TestHelper> basicTests = new ArrayList<>();
-    basicTests.add(
-        new Builder()
-        .name("Write and read mandatory Characteristics")
-        .setUp(context, bluetoothDevice, ProtocolV2.CONFIG_SERVICE_UUID, testCallback)
-        .connect()
-        .insertActions(writeAndReadMandatoryCharacteristics())
-        .disconnect()
-        .build()
-    );
-    basicTests.add(
-        new Builder()
-            .name("Has valid Advertisement Packet")
-            .setUp(context, bluetoothDevice, ProtocolV2.CONFIG_SERVICE_UUID, testCallback)
-            .checkAdvPacket()
-            .build()
-    );
-    basicTests.add(
-        new Builder()
-            .name("Values that are written are advertised")
-            .setUp(context, bluetoothDevice, ProtocolV2.CONFIG_SERVICE_UUID, testCallback)
-            .assertAdvFlags(TestData.BASIC_GENERAL_DATA[0])
-            .assertAdvTxPower(TestData.BASIC_TX_POWER[1])
-            .assertAdvUri(TestData.BASIC_GENERAL_DATA)
-            .build()
-    );
-    return basicTests;
-  }
   private static ArrayList<TestHelper> lockStateTests(Context context,
       BluetoothDevice bluetoothDevice, TestCallback testCallback) {
     ArrayList<TestHelper> lockTests = new ArrayList<>();
@@ -288,14 +255,5 @@ public class UriBeaconTests {
         .write(ProtocolV2.POWER_MODE, TestData.BASIC_GENERAL_DATA, expectedReturnCode)
         .write(ProtocolV2.PERIOD, TestData.BASIC_PERIOD, expectedReturnCode);
   }
-  private static Builder writeAndReadMandatoryCharacteristics() {
-    return new Builder()
-        .write(ProtocolV2.RESET, TestData.BASIC_GENERAL_DATA, BluetoothGatt.GATT_SUCCESS)
-        .assertEquals(ProtocolV2.LOCK_STATE, TestData.UNLOCKED_STATE, BluetoothGatt.GATT_SUCCESS)
-        .writeAndRead(ProtocolV2.DATA, TestData.BASIC_GENERAL_DATA)
-        .writeAndRead(ProtocolV2.FLAGS, TestData.BASIC_GENERAL_DATA)
-        .writeAndRead(ProtocolV2.POWER_LEVELS, TestData.BASIC_TX_POWER)
-        .writeAndRead(ProtocolV2.POWER_MODE, TestData.BASIC_GENERAL_DATA)
-        .writeAndRead(ProtocolV2.PERIOD, TestData.BASIC_PERIOD);
-  }
+
 }
